+++ bin/handler-graphite-status.rb
@@ -4,26 +4,35 @@
 # for details
 #
 # This will send the check status (0,1,2,3) to a graphite metric when a check event state changes
-# Based on handler-graphite-notify.rb
-# See http://imansson.wordpress.com/2012/11/26/why-sensu-is-a-monitoring-router-some-cool-handlers/
+# Based on handler-graphite-notify.rb.
+# Config file by default is graphite_status.json, but can be called with a specific config file
+# using the -j option. This allows multiple graphite handlers to be configured.
 
 require 'sensu-handler'
 require 'simple-graphite'
 
 class Resolve < Sensu::Handler
+  option :json_config,
+         description: 'Config Name',
+         short: '-j JsonConfig',
+         long: '--json_config JsonConfig',
+         required: false,
+         default: 'graphite_status'
+
   def handle
-    port = settings['graphite_notify']['port'] ? settings['graphite_notify']['port'].to_s : '2003'
-    graphite = Graphite.new(host: settings['graphite_notify']['host'], port: port)
+    json_config = config[:json_config]
+    port = settings[json_config]['port'] ? settings[json_config]['port'].to_s : '2003'
+    graphite = Graphite.new(host: settings[json_config]['host'], port: port)
     return unless graphite
     prop = @event['check']['status']
-    message = "#{settings['graphite_notify']['prefix']}.#{@event['client']['name'].tr('.', '_')}.#{@event['check']['name']}"
+    message = "#{settings[json_config]['prefix']}.#{@event['client']['name'].tr('.', '_')}.#{@event['check']['name']}"
     message += " #{prop} #{graphite.time_now + rand(100)}"
     begin
       graphite.push_to_graphite do |graphite_socket|
         graphite_socket.puts message
       end
     rescue ETIMEDOUT
-      error_msg = "Can't connect to #{settings['graphite_notify']['host']}:#{port} and send message #{message}'"
+      error_msg = "Can't connect to #{settings[json_config]['host']}:#{port} and send message #{message}'"
       raise error_msg
     end
   end