+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Added
 - Added handler-graphite-status.rb to create metric which gives the status of the sensu check when it changes state (0,1,2,3)
 
+### Changed
+- Added option -j to handler-graphite-notify.rb to allow configuration file to be defined on command line. This matches the 
+  config in handler-graphite-status.rb
+
 ### Fixed
 - Correct error message for incorrect graphite expression